backend/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==2.2
django-cors-headers==3.5.0
django-storages==1.10.1
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
dropbox==10.8.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
zipp==3.4.0
psycopg2==2.8.6
whitenoise==5.2.0
django-rest-swagger==2.2.0